+++ benchopt/utils/files.py
@@ -1,5 +1,11 @@
-from pathlib import Path
 import warnings
+from pathlib import Path
+from itertools import takewhile
+
+import pandas as pd
+
+from ..config import DEBUG
+from .shell_cmd import _run_shell
 
 
 def rm_folder(folder):
@@ -14,7 +20,7 @@ def rm_folder(folder):
 
 
 def uniquify_results(file_path):
-    "Add a number to filename if it already exists"
+    "Add a number to filename if it already exists."
     if file_path.exists():
         parent = file_path.parent
         stem = file_path.stem
@@ -25,7 +31,30 @@ def uniquify_results(file_path):
         alternative = parent / f"{stem}_{i}{suffix}"
         warnings.warn(
             f"{file_path} already exists. Saving results to {alternative}"
-            )
+        )
         return alternative
     else:
         return file_path
+
+
+def read_results(file_path):
+    """Read benchmark results as pandas DataFrame."""
+    with open(file_path, "r") as f:
+        metadata = {}
+        for line in takewhile(lambda s: s.startswith("#"), f):
+            k, v = line.split(':')
+            metadata[k.strip('#')] = v.strip('\n')
+        df = pd.read_csv(f, comment='#')
+    return df, metadata
+
+
+def write_results(df, file_path):
+    """Write benchmark results to CSV file."""
+    err, tag = _run_shell("git describe --tags --abbrev=0", return_output=True)
+    if err != 0:
+        if DEBUG:
+            print(err, tag)
+        tag = None
+    with open(file_path, 'w') as f:
+        f.write(f'# benchmark-git-tag: {tag}\n')
+        df.to_csv(f)